Move functions that change ScreenBase into a separate file.
