+++ drape_frontend/animation/scale_animation.hpp
@@ -10,7 +10,7 @@ class MapScaleAnimation : public Animation
 {
 public:
   MapScaleAnimation(double startScale, double endScale,
-                    m2::PointD const & globalPosition, m2::PointD const & offset);
+                    m2::PointD const & globalScaleCenter, m2::PointD const & pixelCenterOffset);
 
   Animation::Type GetType() const override { return Animation::MapScale; }
 
@@ -38,8 +38,8 @@ public:
 
 private:
   ScaleInterpolator m_scaleInterpolator;
-  m2::PointD const m_pixelOffset;
-  m2::PointD const m_globalPosition;
+  m2::PointD const m_pixelCenterOffset;
+  m2::PointD const m_globalScaleCenter;
   TObjectProperties m_properties;
   TAnimObjects m_objects;
 };
